APPDEV-801: Holiday Planning
git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/trunk@10967 842219ff-9341-0410-a943-573460bdd4de

+++ projects/TransportationAdmin/src/com/freshdirect/transadmin/web/json/DispatchProviderController.java
@@ -571,7 +571,8 @@ public class DispatchProviderController extends JsonRpcController implements
 				ScribLabel _tempLabel = dispatchManagerService
 						.getScribLabelByDate(TransStringUtil
 								.getServerDate(date));
-				getDispatchManagerService().removeEntityEx(_tempLabel);
+				if(_tempLabel!=null)
+					getDispatchManagerService().removeEntityEx(_tempLabel);
 				ScribLabel _sLabel = new ScribLabel(_scribDate, label);
 
 				return dispatchManagerService.addScribLabel(_sLabel);
